requirements.txt
requests>=2.28.0
psutil>=5.9.0
python-dotenv>=0.20.0
tiktoken>=0.3.0
crawl4ai>=0.1.0